+++ sonar-css-plugin/src/main/java/org/sonar/css/plugin/CssPlugin.java
@@ -22,7 +22,8 @@ package org.sonar.css.plugin;
 import org.sonar.api.Plugin;
 import org.sonar.api.config.PropertyDefinition;
 import org.sonar.api.resources.Qualifiers;
-import org.sonar.css.plugin.bundle.CssAnalyzerBundle;
+import org.sonar.css.plugin.server.bundle.CssAnalyzerBundle;
+import org.sonar.css.plugin.metrics.MetricSensor;
 import org.sonar.css.plugin.server.CssAnalyzerBridgeServer;
 
 public class CssPlugin implements Plugin {
