+++ sonar-css-plugin/src/main/java/org/sonar/css/plugin/metrics/CssTokenType.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.css.plugin;
+package org.sonar.css.plugin.metrics;
 
 import com.sonar.sslr.api.AstNode;
 import com.sonar.sslr.api.TokenType;
