+++ sonar-css-plugin/src/main/java/org/sonar/css/plugin/server/bundle/package-info.java
@@ -18,4 +18,4 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 @javax.annotation.ParametersAreNonnullByDefault
-package org.sonar.css.plugin.bundle;
+package org.sonar.css.plugin.server.bundle;
