+++ sonar-css-plugin/src/test/java/org/sonar/css/plugin/metrics/TokenizerTest.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.css.plugin;
+package org.sonar.css.plugin.metrics;
 
 import java.util.List;
 import org.apache.commons.lang.StringUtils;
